Moved CTA's to Body, planning to create proper footer
I want to build an exploded footer as my primary navigation (for now). I want to build the site on my own time, and need to do that without the primary nav. Hence, footer.

To do it properly, need to move the CTA's to the body so I can work on the footer-proper.