Working on getting everything set for /journal/ ot start going live.
this is focused on the journal blueprint, some typographic bits, and better html structure. also working on the blueprint for file uploads so I can add alt-text to images uploaded in markdown area.../image.yml didn't work but /default.yml did. need to figure out what's up there...but that's for another day.