refactored and optimized quite a bit
theme picker: removed the scroll to top stuff, combined with theme-picker things, and added the floating buttons for both theme picker and the scroll to top. optimized and fixed from there.

the table of contents: made the buttons work more like i want with the animations

work-landing: added the .yml so i can edit things via the cms rather than through the html.

in the theme selection refactoring, i adjusted hte data-theme side of things on the :root so now it's clearer. also removed the flashing of the light and dark mdoes.

added a "sticky bottom" sort of thing, so that the floating buttons for theme and scroll-to-top disappear and there's a section in the footer for it.

also adjusted footer background colors, made better.